Handle numerical failures in velocity dispersion calculation
In the generic dark matter density profile class when computing velocity dispersions the numerical integration can sometimes fail to reach the required tolerence. This is usually in extreme cases (e.g. a subhalo close to being tidally disrupted). It is useful to not fail in these cases, but to instead relax the tolerance and try again. This patch implements such a "backing off" of the tolerance. If the integral still does not converge at a relatively low tolerance it will then fail with an appropriate error message.