Move dark matter profile scale evolution into a nodeOperator
Moves interpolation of dark matter profile scale radii out of the darkMatterProfile component and into a nodeOperator.

Adds a "concentrationLimiter" dark matter profile scale radius setting class which limits the scale radius such that the concentration lies within a specified range. This removes the application of this limit within the darkMatterProfile component itself. Also, this obsoletes the need for the "scaleIsLimited" property of the darkMatterProfile component.

Also removes the "preset" darkMatterProfile class as it is now obsolete.

Cleans up unused nodeOperator sections in various test suite parameter files.

Allows a nodeOperator to be applied to a tree in the "read" mergerTreeConstructor after the tree is established (i.e. times, masses, and pointers are set up), but prior to setting any additional properties of the tree (e.g. orbits). This allows setting of some properties (e.g. scale lengths) that may be required prior to setting other properties using data read from file (e.g. orbits).