Improvements to the "energy tensor" N-body operator.
The measurement is now made in a frame in which the net linear and angular momentum of the particles is zero. Additionally, after measuring the kinetic and potential energyt tensors the total energy tensor is constructed, its eigenvectors are found, and the coordinate basis is rotated into alignment with these eigenvectors. The kinetic and potential energies in this coordinate basis are then computed and output.