Avoid tabulating mean halo density to too early times
Previously the mean halo density was tabulated to a time t/10 when the result was requested at time t. While this ensures that tabulated results for much earlier times are readily available it can result in problems in some cases. Specifically, when using filtering mass calculations with initial conditions set using the Gnedin results it is possible to exceed the earliest valid time for these. This patch changes the tabulation range to t/2 which mitigates (but does not completely resolve) this issue.