Improve parallelism of the test suite
Enables all test suite tests to be submitted simultaneously. All required code is first built, after which all jobs which handle their own submission to PBS are launched in a new thread, after which all PBS jobs are submitted simultaneously. Once PBS jobs are finished the thread is joined and testing is done.

Also adds a test job which compiles documentation and checks for errors.