Ensure correct times are assigned to star formation history
When tabulating SEDs, for the first snapshot we need to set the time of the ndoe to slightly before the first snapshot time to ensure that star formation histories are assigned times appropriate to that snapshot. Otherwise, if the node is placed exactly at the first snapshot time, the times assigned to star formation histories will be those corresponding to the next snapshot (since the first snapshot will be considered to have already been reached).