fix: Pass "node" object to critical overdensity functions to allow environment to be accessed
The "environmental" critical overdensity class must be able to access the halo environment, which requires it to be passed a treeNode object, "node", which has a pointer to the host tree (which stores the environment). This fix ensures that happens for several calling functions, and adds an error message from the "environmental" critical overdensity class is "node" is not present.