fix: Avoid retabulation of delta_crit(t) during merger tree building
If delta_crit(t) gets retabulated during merger tree construction it can lead to branch mis-ordering which leads to an infinite loop. A temporary fix is to force delta_crit(t) to be pre-tabulated to the earliest time at which it will be needed. This fix extends that pre-tabulation to slightly earlier times still to avoid any retabulation being triggered by small oversteps of the earliest time.