fix(IO): Improve HDF5 dataset chunking
When creating appendable, rank > 1 arrays with HDF5 we previously always set the chunk size to the supplied value in the first dimension, and set chunksize in all other dimensions to 1. This is very inefficient and can lead to growing memory usage when doing many appends to a dataset. This patch changes the approach such that we now use the supplied value to set the chunk size in the dimension to which we will append, and uses the actual dataset size as the chunk size in all other dimensions.