fix: Re-use tabulated excursion set solutions
The final check on whether the tabulated and stored-to-file excursion set solution is sufficient was made only if the table was not already initialized. This was inefficient as if a retabulation was required the stored-to-file solution would never be checked, resulting in duplication of effort between parallel threads. This patch causes the check of the stored-to-file solution to be made even if the table has already been initialized.