fix: Covariance model for volume functions and cross-correlation term for mean functions
This patch corrects the calculation of the covariance contributions from main branch halos when constructing volume functions. The approach now has the correct covariances under the assumption of a multinomial distribution for the contributions to the bins of the function.

Additionally, this patch removes the "hack" of computing the cross-correlation between weighted and unweighted functions needed when finding the covariance in mean functions. The hack was to evaluate a volume function with weight multiplied by the square root of the weighting property. This works for cases where the weighting property is always positive, but fails for negative values. The new approach computes the covariance directly, and correctly handles negative values.

Finally, the square root property operator is removed as it is no longer needed.