fix: "boolean" property operator should never return -1
The "boolean" property operator is used to allow filtering out of nodes when computing mean functions. It should return zero if the property has been set to zero (indicating that the node should be filtered out), and 1 otherwise (so that the accumulated weight is just the regular node weight). Previously this class would return -1 if the property was negative.