feat: For progenitor mass function analyses allow single-parent-mass and variable root-variance
Single parent mass cases are now possible providing that minimum and maximum parent masses are specified. (Also allows for non-uniform parent mass bins in the multi-parent mass case.)

If using a "rootVarianceTargetFractional" this can now be set as an array allowing different values to be used for each bin of progenitor mass ratio.