refactor!: Replace "interoutput" node component with node operator/extractor
Removes the "interoutput" node component and replaces it with a node operator which computes the mean rates of star formation in disk and spheroid between successive outputs along with a node property extractor to extract (and output) these rates.

BREAKING CHANGE: changes both the input parameter names and output property names for these rates.