fix(memory): Fix memory leak when calling root finder with error status
If the root finder failed and an error status argument was present, the "currentFinderIndex" was not decremented before returning. This could lead to excessive growth of the finder stack and substantial performance degradation as the stack was frequently extended (getting slower each time as the stack grew in size). The counter is now correctly decremented in these cases.