feat: Add reporting on total memory used
Detects if the Linux /proc filesystem is available. If it is, /proc/self/stat is used to find the current resident set size when reporting memory usage. This is used to find the amount of memory allocated to "unknown" objects, which is then included in the memory report. This should allow for more realistic reports of total memory used even in cases where memory leaks are present.