fix: Make "normal" halo environment more consistent
The distribution function of overdensities is now a true PDF (i.e. is normalized to 1).

A new function is added to the haloEnvironment class which returns the fraction of the volume occupied by the environments described by the class.

The "environmentAveraged" halo mass function class now correctly incorporates this volume fraction (unless instructed not too).

A test code is included to verify that the environment-averaged mass function matches the unconditioned mass function for the case of the Press-Schechter mass function.