fix: Add tolerances for minimum metallicity in disk and spheroid components
Adds parameters which control the minimum metallicity (for gas and stars) in the standard disk and spheroid components. Prior to this, galaxies with very low metallicity could be rate-limited by that metallicity in ODE solver steps. Since we likely do not care about such tiny metallicities it makes sense to introduce some tolerance.

Also removes the factor 0.1 introduced into the absolute tolerance scale for gas mass in the standard spheroid component - this was introduced long ago and seems to be no longer needed, so it makes sense to restore parity with the treatment in the standard disk component.