fix: Refactor the satellite nodeComponent
Stage 1 of this refactoring. Replace the "verySimple" satellite component with a "mergeTime" implementation. Move the functionality that existed in "verySimple" into a nodeOperator - the "mergeTime" nodeComponentSatellite now simply implements the component and a single "timeOfMerging" property.

The dual properties "timeOfMerging" and "timeUntilMerging" are now replaced by just a single "timeOfMerging" throughout all nodeComponentSatellite implementations. This means we no longer have to evolve "timeUntilMerging" as part of the ODE system.