fix: Set correct merging time for instantaneously merging subhalos
When reading merger trees from file, in cases where a non-primary progenitor has no subhalo descendent we assume instantaneous merging of the halo. Previously the merging time in such cases was set to the time at which the halo existed, instead of the time at which its parent exists (which is when the node merger actually occurs).