fix: Use consistent `darkMatterProfileClass` object for galactic structure calculations
The `galacticStructureStandard` class uses a shortcut when computing the radius enclosing a given mass if the mass requested is dark matter only. Previously this used a different `darkMatterProfileClass` object for this calculation then for corresponding calculations of the mass within a given radius. In some instances this could lead to inconsistent results.

This patch avoids this by using the same `darkMatterProfileClass` object for both calculations.