fix: Correctly handle unnormalized transfer functions
In the transferFunctionEnvelope class, when computing mode masses we must account for the fact that the transfer functions used may not share the same normalization as k->0. This fix causes the transfer functions to be used to be normalized to T(k)=1 at small k.

Additionally, this patch allows the enveloping to be applied to the ratio of the transfer function and the reference transfer function (instead of to the transfer function directly).