fix: Test for existance of solution prior to use
When solving for equlibrium radii of galactic structure, on the first iteration we use the dark matter-only profile to find an initial guess. In some instances (e.g. strongly heated halos) it is possible that no solution exists. Check for this and, if it is the case, then simply use the virial radius as an initial guess instead.