fix: Add a workaround for missing `pthread_mutex_destroy()` function in `GLIBC >= 2.34`
In `GLIBC >= 2.34` the functionality `libpthread` is absorbed into `libc`. This results in statically-linked Fortran codes not having `pthread_mutex_destroy()` linked into them. This adds a workaround by including a dummy C function that calls `pthread_mutex_destroy()` (but which itself is never called) for force `pthread_mutex_destroy()` to be linked.