fix: Avoid network file system sync issues
This attempts to fix sync issues with network file systems which were causing erroneous results from the `File_Exists()` function (returning `true` even though the file did not exist).

To attempt to mitigate this we run an `opendir()` followed by a `closedir()` on the parent directory prior to the file existance check, which invalidates any network file system cache.