fix(perf): Optimize event hook code
Instead of checking each hooked event on each call to see if it matches the appropriate OpenMP thread(s), we now instead perform this check once, at the start of each new OpenMP parallel region, creating a custom array of functions which need to be called by that thread. Execution of the event then simply involves iterating through that array and calling each function, with no checking to be done. The parent thread's array is restored on exiting the OpenMP parallel region.