fix: Avoid sync on file unlock
In the `intergalacticMediumFilteringMassGnedin2000` class we attempt to read a tabulated solution, obtaining a read lock on the file before, and releasing it after. If the file does not exist the unlock will create a zero-size file unless we disable sync. Creation of a zero-size file can lead to crashes if other threads attempt to read it. So, sync is now disabled - there should be no need to sync after a pure read access anyway.