fix: Prevent the `mergerTreeEvolveTimestepHostTidalMassLoss` from taking tiny timesteps
If tree evolution is using the `backtrackToSatellites` option, this timestep limiter could lead to extremely tiny timesteps for a satellite as it is allowed to evolve only a slight amount beyond the time of its host. This fix prevents this by refusing the evolve the satellite (if this class is the limiting factor in evolution) unless the step that can be taken is a sufficiently large fraction of the mass loss timescale. If a step is refused, this will force the host halo to evolve instead. This evoids tiny timesteps.