fix: Set `pipefail` in CI/CD to catch exit status of piped processes
In CI/CD we pipe the ouptut of tests through `tee` so that a log is recorded. But, this means that the exit status is that of `tee`, which is always `0` even if the test code fails (i.e. has a non-zero exit status). By setting `pipefail` we get the last non-zero exit status (or zero if all exit statuses in the pipeline were zero). This ensure that we catch test code failures.