fix: Make integration of σ(M) more robust
Adds a further split in the integral for σ(M) at a wavenumber corresponding to a few times the inverse of the top-hat radius. Since window functions are expected to cut off above this wavenumber this can help to capture the peak in the integrand (which, otherwise, was very occassionally not captured). As this results is multiple different wavenumber splits for these integrals, we now handle this in a more general way, first generating a list of wavenumbers at which to split, sorting these, and integrating each resulting interval individually.