fix: Avoid root search range failure in `font2008` hot halo ram pressure stripping class
Prevoiusly upper and lower limits on the root interval were tested and, if a root existed in the interval, these were passed to the root-finding function. But, due to retabulations and the associated rounding errors it was possible on rare occasions for the root to move slightly outside of the interval between testing and calling the root finder. To avoid this we now add some buffer to the interval passed to the root finder.