fix: Allow `nbodyImporterGadgetBinary` class to handle files with larger numbers of particles
Gadget binary files are written in Fortran unformatted form. This should mean that we can access them using the standard, record-block based approach. However, Fortran (specifically gfortran) uses a 32-bit integer to store the prefix and suffix record size for each record
block. Integers in Fortran are always signed. But, often these Gadget files have been written by C code, with an unsigned int used for the record size prefix and suffix. This can cause problems, as the record size (which would be a valid number if interpretted as an unsigned int) is read by Fortran as a negative number, leading to failed reads.

To avoid this, we use stream access, read the record size prefix and suffix directly, and then ignore it - assuming that we know the size of the data in each record (which we do, providing that the Gadget format has been followed correctly).