fix: Avoid unnecessary divide/multiply in summation cooling function
When computing the logarithmic gradient of the cooling function with respect to density in the `coolingFunctionSummation` class we were summing logarithmic gradients divided by the hydrogen density, and then multiply by this after summation. But, since this density is constant there's no need to do this, and we then also avoid floating point errors if the density is zero.