fix: Improvement to internal I/O thread safety
This fix automates the application of OpenMP `gfortranInteralIO` critical sections around internal (and external) I/O when the `-DTHREADSAFEIO` compile option is specified. It also translates (in the preprocessor) `FoX_DOM_Access` critical sections to `gfortranInternalIO` critical sections when the `-DTHREADSAFEIO` compile option is specified since some internal I/O occurs in the FoX library also. (Note that this still ensures that the FoX library is accessed only by a single thread at a time, but now strengthens that protecting to avoid conflicts with other threads performing internal I/O).