fix: Imrpove run-time scaling of `allowedParameters()` methods for very large classes
Previously this worked with a simple approach of appending parameter names to a list, with no checking if those names were already on the list. This involves extending the list to allow new entries to be added, which in turn involves (potentially many) cycles of array creation, copying, and destruction. For large classes (e.g. ones that composite hundreds to thousands of other classes) this can be slow as it scales as O(N^2). This new approach only adds new entries if they are not already on the list. So, it has higher overhead for short lists (but that is not really a problem), but avoids the list becoming huge which resulted in the poor scaling.