fix: Use a linked-list of `galacticStructureStandard` state
In the `galacticStructureStandard` class we make use of a stack to store the state of the query (e.g. radius, mass and component types, etc.). This is necessary as the state must be accessible with sub-module scope (for some of the numerical solvers used), and these functions can be called recursively with different state. So, we must push and pop from a stack in order to avoid overwriting state from previous invocations.

This was previously implemented as an allocatable array of states, which was expanded as needed. This can cause problems though. If variables from the state stack have been passed (by reference) to some function, which then invokes the `galacticStructureStandard` class itself, and if that causes the stack to be expanded, those variables passed by reference now point to memory is no longer in use, and so it is possible for that memory to be over-written, leading to undefined behavior.

To avoid this, we now use a linked list for our stack. This can be extended by simply allocating a new entry at the end of the list, without moving the memory associated with the earlier parts of the list.