fix: Use a counter for GSL error reporting state
This replaces the previous boolean state which would fail if nested functions switched GSL error handling off/on (this could leave the error handler on when the outer function expected it to be off). Using a counter avoids this problem - default error handling occurs only when the counter is at zero.