fix: Call the decorated `mergerTreeBuildController` `conrtol()` method for single-step control
This is necessary to allow the decorated controller to do whatever it may need to do. We then simply override its results if necessary to force a single step.

Also adds an option to not halt after a single step - this is useful to allow a fixed size step to be taken, and then have the tree build continue as normal.