feat: Improve extrapolation in stellar parameters
In the `stellarAstrophysicsFile` class, extrapolation in initial stellar mass is no longer performed - instead, for masses outside of the available range, stellar properties are truncated to zero. Additionally, no extrapolation in metallicity is made outside of the available range. Instead metallicities outside of the available range are pinned the the minimum/maximum metallicity available. This avoids unphysical extrapolations.

Additionally, expansion of file paths is enabled for convenience.