fix: Correct the implementation of the number of Type Ia SNe
This patch corrects the implementation of the Nagashima et al. (2005) model for the number of Type Ia SNe as a function of population age/metallicity. It correct the dependence on the initial mass function, adds a missing Jacobian term, and implements the upper limit on the primary mass. A new test is included to compare this to results from Nagashima et al. (2005).