fix: Ensure that adaptive star formation histories are always created with the correct length
Previously, if a spheroid was created via interrupt at precisely the next output time, the star formation history (created by the `starFormationHistoryAdaptive` class) would be created with times appropriate to the subsequent output time. This resulted in mismatches between the lengths and times of star formation histories. To avoid this, when created via an interrupt, the current target output time is provided, allowing the correct set of star formation history times to be assigned in such cases.