feat: Expand the test of adaptive star formation histories
Now additionally checks the lengths of spheroid star formation histories are identical, checks that disk and spheroid star formation histories are of the same lengths, and that the times at which the star formation histories are tabulated are consistent with the output time. Furthermore, star formation histories are output for the "all" component to verify that this functionality works as expected.