fix: Fix uninitialized content in output of filter properties
When outputting filter information (names and effective wavelengths), we previously output the entirety of the allocated array, even if some parts of the array were unpopulated. This can happen as the array is dynamically expanded in size by a factor 2 as needed.

Additionally, this patch ensures that effective wavelengths for all filters have actually been computed before attempting to output them.