fix: Fix static builds of FSPS and CLASS
For FSPS, the `Makefile` syntax had changed meaning that our patch for force static builds failed. This is now fixed and we include some level of checking that our patch applied correctly.

For CLASS, the `-static` flag was not being correctly applied to the `LDFLAGS` variable used when linking the code.

The CI/CD workflow now includes checks that these codes are statically linked.