fix: `satellite` components must be back-propagated during node promotion
This is necessary as some operators cause orbits to be assigned to non-primary progenitor halos. Without back-propagation of the `satellite` component those orbits (and other satellite properties) will be lost when the primary progenitor of the halo is promoted to it (and, typically, a new orbit is then assigned, which will break the expected correlations between subhalo orbits and any properties of the host halo that depend on them).