fix: Fix solution to the numerical Jeans integral for spehrical mass distributions
When numerically integrating the Jeans equation in spherical mass distributions we must make sure that we begin at a sufficiently large radius that the contribution from outside of the radius is negligble. This is now achieved by seeking an outer radius at which the density is far below theat at the largest radius of interest in the tabualted solution.