fix: Avoid segfault when closing HDF5 during error
When an error occurs we attempt to close down the HDF5 subsystem. This requires obtaining a lock on the HDF5 subsysten. But, if an error occurs before the lock was initialized then this could cause a segfault due to attempting to access the uninitialized lock. We now check the initialization status before attempting to lock. This could potentially result in a race condition, but since this is already occuring during error, there's only so much that we can do.