fix: Avoid deadlocks when reporting the location of the `datasets` repo
Previously, the `inputPath()` function was called within a `FoX_Access` OpenMP `critical` section. If compiled with `-DTHREADSAFEIO` such `critical` sections are changed into `gfortranInternalIO` `critical` sections. This then caused a deadlock with the same `critical` section within the `displayMessage()` function when diplaying the location of the `datasets` repo. To fix this, the `inputPath()` function is now called prior to entering the `FoX_Access` OpenMP `critical` section.