feat: Improve memory reporting when running under SLURM
If SLURM has set memory limits (and reports them in its environment variables) we now use those to set the maximum available memory when reporting memory usage. A `NOTE` is displayed to show whether the total memory available was determined from SLURM or from the system.

Also, when running under MPI, the maximum available memory is divided by the number of processes on the node to give an indication of how much is available to each process.