feat: Add surface density methods and improve `subset()`
Provides methods to find the radius enclosing a surface density, and to report if a monotonically-decreasing surface density can be assumed.

Improves the `subset()` method to return the object itself if the subset contains only a single object (instead of wrapping it inside a `massDistributionComposite` object. This enables special behavior for certain classes (e.g. if an analytic solution is available for an exponential disk).