feat: Provide more informative error messages when failing to parse radius specifiers
The problem radius specifier is displayed along with an error message explaining the problem, and the problematic part of the specifier highlighted in red. Furthermore, in cases where the problem is an unmatched enumeration member, the allowed enumeration members are listed. Lastly, a link to the relevant section of the documentation is provided.