fix: Set the vector angular momentum (if possible) in the `nodeOperatorHaloAngularMomentumRandom` class
The vector angular momentum is assumed to always point along the first axis (since this model makes no physical prediction for the direction of the vector).

Setting the vector angular momentum is important to ensure that the angular momentum evolves correctly (since the evolution in the magnitude of the angular momentum is determined from the evolution in the vector components).