fix: Fix indexing of nodes in parallel tree builds
We must use a "global" `nodeIndex`. Using a stack-local `nodeIndex` means that the value is frozen for each task at the time that the task is created. This results in reuse of node index values.

Instead, we now store `nodeIndex` in `self` so that it is shared between all threads.

Additionally, `hostTree` pointers in newly-created `treeNode`s are now set using the `hostTree` pointer from the branch tip node. This avoids a problem (possibly a compiler bug?) when using `tree` directly that caused the `hostTree` pointer to be pointed at seemingly random memory addresses.