fix: Correctly set the final time for the star formation history
In the `starFormationHistoryAdaptive` class the final bin in the SFH can extend past the corresponding output time (due to the finite minimum width of the time bins). When returning the set of times at whicht he SFH is tabulated, limit this final bin to the output time - this ensures that we have the most accurate estimate of the SFR in this bin. Without doing this, the size of the bin is effectively overestimated (since it can be populated with star formation only up to the output time) which would result in an underestimate of the SFR.