fix: Position interpolation was broken - fix it
Commit edc8a2cd1c8c43578c29909fa8dda346ce406ed2 broke interpolation of N-body positions because:

1. Any set position was ignored in favor of a position drawn from the `positionHistory` if available;

2. The `nodeOperatorPositionInterpolated` class was computing the interpolation but never actually setting the position.

Additionally, lightcone classes, which rely on this interpolation to find the exact time of lightcone crossing, were no longer getting interpolated position as they simply updated the `basis%time()` property. This used to work, but with the refactoring in commit edc8a2cd1c8c43578c29909fa8dda346ce406ed2 we instead need to call the `solveAnalytics` method of the `nodeOperator` to set the positions.