feat: Optimize `starFormationRateSurfaceDensityDisksBlitz2006` table construction
Previously, objects kept independent copies of this table and never tried to re-read the version from file. This meant that, if the file had been updated, an object would ignore it, recompute the table itself, and write it out. This lead to inefficiency as the new table might not have as large extent as the version it overwrote.

Now, the file is re-read whenever the current table owned by an object is insufficient.