feat: Track the last modified commit hash in parameter files
The `version` element in parameter files is now replaced with a `lastModified` element that tracks both the time at which the file was last modified, and the `git` commit hash for which it was intended. At run time, this is now used to check if the parameter file is out of date (i.e. is missing parameter migrations), or if the executable is out of date.

To facilitate this, the parameter migration data is no longer embedded in the `parametersMigrate.pl` script, but is stored in its own XML file.