fix: Avoid mass profiles that increase inward
For poorly-determined mass profiles we could occasionally get cases where the innermost point had a higher mass than the next outer point. When extrapolated this leads to a mass that increases inward, which is unphysical and leads to failures. When working with such poorly-determined profiles this situation is now tested for and avoided.