fix: Use an efficient algorithm for detecting branches off of the main branch
When retaining subsampled stubs on the main branch only we need to test whether a branch is off of the main branch. Generic main branch detection is slow as it can involve walking up through the entire tree. This fix introduces an optimization by storing the unique ID of the likely next-visited node on the main branch which we can then use to quickly detect that a branch is off of the main branch.