feat(perf): Add an efficient main branch filtering `nodeOperator`
Implements a `nodeOperator` that filters another `nodeOperator` by applying it only to main branch nodes (or, optionally, non-main branch nodes) during node tree initialization. It makes use of the fact that this initialization occurs in a depth-first walk to more efficiently determine if a node is or is not on the main branch.