fix(perf): Allow the gravitational potential integral to use a change of variables
The integral for gravitational potential:
$$
\int -\frac{\mathrm{G}M(r)}{r^2} \mathrm{d}r,
$$
can now used a change of variable to allow integration over some other radius coordinate. This is exploited in heated dark matter profiles to integrate over the initial radius (instead of the final radius) which allows for much more rapid evaluation of the integrand (as we no longer have to use a root finder to determine the initial radius given the final radius).