feat: Add an option to set dark matter profile scale radii on mass increases by a given factor looking "backwards" along each branch
Normally, if the scale radius is to be update when the halo mass increases by a factor of, for example, 2, we start from the tip of the branch, set a scale radius there, and walk forwards up the branch until the mass is increased by this factor. An alternative is introduced in which the scale radius is always set for the end fo the branch (just before it mergers), and propagated backward until the mass of trhe branch is decreased by a factor 2. This can be useful to ensure that the scale radius is correct at the time of branch merging.