fix: Add a trailing newline to the last line of source files if not present when preprocessing
Some editors do not add a newline to the last line of the file by default. The lack of a newline can cause problems with parsing the file as this last line may be concatenated with further code - with no newline the concatenated code will be erroneously appended to the last line.