fix: Prevent periodic wrapping of interpolated halo positions
Adds
```
<wrapPeriodic value="false"/>
```
for the position interpolation. This prevents it wrapping halos that move outside of the original simulation box as they are interpolation back into the other side of the periodic box. Doing such a wrap makes it look like the halo moved across the entire box volume instantaneously, resulting in spurious lightcone crossings. With this option, the halo is instead left outside of the original periodic volume. Since we tesselate the volume when constructing the lightcone this isn't a problem - the halo will simply appear in the adjacent replication of the box.