fix: Allow some tolerance in when we process a lightcone crossing
Allow some tolerance in the time at which a lightcone crossing halo is output.This is necessary as a node's evolution could be stopped very close to (but not quite at) the crossing time, and in some cases the crossing time would then not be rediscovered on the next time step (due to numerical precision). Using an absolute tolerance of 10⁻⁹ Gyr means that we at most get the position of the galaxy on the lightcone wrong by 1 light year - a tiny amount.