feat: Allow the primary output group to be changed from the root group
Normally, the `Outputs` group (and all sub-content) are written directly to the root group of the output file. This commit introduces functionality to allow this to be changed such that the `Outputs` group can be created in some other group. This allows for multiple models (and their `Outputs`) to be stored in a single file, for example.