feat: Disallow the combination `[setOutputGroup]=true` and `[collaborativeMPI]=true`
`[setOutputGroup]=true` results in the results of each model evaluation written to its own group, but `[collaborativeMPI]=true` results in each MPI process evolving a subset of trees from each model evaluation, and writing them to its own output file - this will result in a random mix of trees in each output group.

A help message is provided when this error is triggered.