fix: Handle edge case for orbital angles read from merger tree files
When constructing subhalo orbits for halos read from merger tree files, if the orbital position vecctor lies prrecisely along the $z$-axis we can not use the normal approach of $\mathbf{r} \times \mathbf{\hat{e}}_\mathrm{z}$ to construct a vector from which to measure the angle $\epsilon$ (angle of motion in the tangential plane). In such cases we now set t$\epsilopn=\pi/2$ which will be consistent with the result obtained at some slightly later time, as now explained in the `keplerOrbit` class.