fix: Use the correct mass distribution when undecorating kinematics distributions
When computing the velocity dispersion in self-gravitating decorated mass distributions, and the undecorated distribution can be used, we were previously undecorating only the embedding mass distribution, not the target mass distribution. This resulted in a mismatch between target and embedding mass distributions, which caused incorrect results and slow-downs (as analytic solutions for self-gravitating systems could not be used).