feat: Allow the `posteriorSampleStateInitializePosteriorMaximumGaussianSphere` class to handle changes in parameter between runs
Parameters from the prior model are now matched to those in the current model by name, rather than by position in the list of parameters. This allows this class to handle cases where the parameters in the current model differ from the old (e.g. having additional parameter present). A fall-back `posteriorSampleStateInitializeClass` object is used to set the initial state of any parameter not found in the prior model output chains.