feat: Set merger tree seeds using a dedicated class
The default implementation usesincrementing seeds based on tree index (matching the prior behavior).

A new `mergerTreeSeedsRandom` implementation chooses seeds based on a non-repeating random sequence. This is intended to allow a simple change by 1 in the seed given in the parameter file to generate a completely new set of trees.