feat: Detect recursion in `objectBuilder` directives
In cases where an object has a dependency on an object of the same class (for example in:
```
  <cosmologicalMassVariance value="scaled">
    <scale value="1.1"/>
    <cosmologicalMassVariance value="filteredPower">
      <sigma_8 value="0.807"/>
    </cosmologicalMassVariance>
  </cosmologicalMassVariance>
```
the `cosmologicalMassVarianceScaled` object depends on another object of the `cosmologicalMassVariance` class), if no instance of the dependency is given then an infinite recursion happens as the object build steps up through the parameter hierarchy looking for a member of the class, and finds the original object, triggering an infinite recursion.

This update checks for such recursion and stops with an error message.